Correct typography according to Devanagari norms
The ASCII vertical bar (|) character (`U+007c`) has special meaning in
many languages and even more so to programming languages and computers.
In spite of its visual similarity is is not an appropriate substitute
for the correct punctuation symbol in Devanagari based languages.
Instead the pūrṇa virām (।) character (`U+0964`) or it's double variant
(॥, `U+0965`) should be used to denote full stops depending on context.

From Wikipedia:

> The end of a sentence or half-verse may be marked with a dot known as
> a pūrṇa virām or a vertical line danda: ।. The end of a full verse may
> be marked with two vertical lines: ॥.